meteor build 增加参数 --allow-superuser
You are attempting to run Meteor as the 'root' superuser. If you are
developing, this is almost certainly *not* what you want to do and will
likely result in
incorrect file permissions. However, if you are running this command in
a build process (CI, etc.), or you are absolutely sure you know what you
are doing,
set the METEOR_ALLOW_SUPERUSER environment variable or pass
--allow-superuser to proceed.

Even with METEOR_ALLOW_SUPERUSER or --allow-superuser, permissions in
your app directory will be incorrect if you ever attempt to perform any
Meteor tasks as
a normal user. If you need to fix your permissions, run the following
command from the root of your project:

sudo chown -Rh <username> .meteor/local